+++ src/decrypt_files.cpp
@@ -239,6 +239,9 @@ bool DecryptStream(
  *          decryption is in progress, it will gracefully terminate
  *          decryption and allow the program to exit.
  *
+ *      force [in]
+ *          If true, the output file will be overwritten if it exists.
+ *
  *      quiet [in]
  *          If true, the program will not emit messages to the terminal, except
  *          for error messages (which are directed to stderr).
@@ -263,6 +266,7 @@ bool DecryptStream(
 bool DecryptFiles(
     Terra::Logger::LoggerPointer parent_logger,
     ProcessControl &process_control,
+    const bool force,
     const bool quiet,
     const SecureU8String &password,
     const std::vector<SecureString> &filenames,
@@ -494,13 +498,13 @@ bool DecryptFiles(
                 // If the output file does not exist, attempt to remove later
                 // (Do not remove by default so as to not attempt to remove
                 // things like character special devices.)
-                if (!std::filesystem::exists(file_status))
+                if (!std::filesystem::exists(file_status) || force)
                 {
                     remove_on_fail = true;
                 }
 
                 // Does a regular file having this output file name exist?
-                if (std::filesystem::is_regular_file(file_status))
+                if (!force && std::filesystem::is_regular_file(file_status))
                 {
                     std::cerr << "Target output file already exists: "
                               << out_file << std::endl;