+++ src/encrypt_files.h
@@ -45,6 +45,9 @@
  *          encryption is in progress, it will gracefully terminate
  *          encryption and allow the program to exit.
  *
+ *      force [in]
+ *          If true, the output file will be overwritten if it exists.
+ *
  *      quiet [in]
  *          If true, the program will not emit messages to the terminal, except
  *          for error messages (which are directed to stderr).
@@ -77,6 +80,7 @@
 bool EncryptFiles(
     Terra::Logger::LoggerPointer parent_logger,
     ProcessControl &process_control,
+    const bool force,
     const bool quiet,
     const SecureU8String &password,
     const std::uint32_t iterations,